requirements.txt
setuptools
wheel
streamlit==1.28.0
pandas==2.0.3
numpy==1.24.3
xgboost==1.7.5
scikit-learn==1.3.0
faiss-cpu==1.7.4
transformers==4.34.0
torch==2.1.0
sentence-transformers==2.2.2
plotly==5.17.0
networkx==3.1
matplotlib==3.7.2
huggingface-hub==0.17.3
protobuf==3.20.3
shap==0.43.0
streamlit-shap==1.0.2
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
faker==19.3.0
openpyxl==3.1.2